Don't clone BonnyCI/zuul into nodepool built images
We needed zuul on nodepool nodes in v2 because the zuul-cloner and such
were activated on the node. In v3 this should all be done by the
executor and then synced onto the worker host. We therefore shouldn't
clone the zuul repo (and the old one at that) onto images.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>